requirements.txt
slack-bolt>=1.18.0
slack-sdk>=3.21.0
python-dotenv>=1.0.0
schedule>=1.2.0
